[feat] 메인 화면_홈 / 위치 최신화 로직 유틸화(#191)
